requirements.txt
Flask==0.11.1
google-cloud==0.22.0
gunicorn==19.6.0
flask-cors==3.0.2
